requirements.txt
python-pptx==0.6.21
groq==0.5.0
flask==3.0.0
pillow==10.4.0
webcolors==1.13
colorthief==0.2.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
werkzeug==3.0.1
